fix: int para long em fatorial iterativo

+++ custo_iteracao_e_recursao/Fatorial.java
@@ -2,7 +2,7 @@ package custo_iteracao_e_recursao;
 
 public class Fatorial {
 
-    public static int fatorialIterativo(int valor) {
+    public static long fatorialIterativo(int valor) {
         int fatorial = 1;
 
         for (int i = valor; i > 0; i--) {
